requirements.txt
asgiref==3.4.1
autopep8==1.6.0
black==21.10b0
click==8.0.3
colorama==0.4.4
Django==3.2.9
djangorestframework==3.12.4
et-xmlfile==1.1.0
mypy-extensions==0.4.3
numpy==1.21.3
openpyxl==3.0.9
pandas==1.3.4
pathspec==0.9.0
platformdirs==2.4.0
pycodestyle==2.8.0
python-dateutil==2.8.2
pytz==2021.3
regex==2021.10.23
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==1.2.2
typing-extensions==3.10.0.2
